backend/requirements.txt
fastapi>=0.133.1
uvicorn[standard]>=0.30.0

google-adk==1.25.1
google-cloud-aiplatform[agent-engines]>=1.132.0

pydantic>=2.11.1,<3
pydantic-settings>=2.1.0

python-multipart>=0.0.9,<0.1
python-dotenv>=1.0.0
structlog>=23.2.0
redis>=5.0.1

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.4
